requirements.txt
pandas
pycantonese
jieba
requests
jupyter
transformers
torchaudio
functools
datasets